rename: common 패키지 세분화
