+++ core/common/src/main/java/com/tenday/core/common/enums/Format.kt
@@ -1,0 +1,5 @@
+package com.tenday.core.common.enums
+
+fun Int.toData(): String {
+    return "%,d".format(this)
+}
