+++ core/common/src/main/java/com/tenday/core/common/extensions/BadgeCode.kt
@@ -1,4 +1,4 @@
-package com.tenday.core.common
+package com.tenday.core.common.extensions
 
 enum class BadgeCode {
     A, B, C, D, E, F;
