+++ core/common/src/main/java/com/tenday/core/common/extensions/JobPosition.kt
@@ -1,4 +1,4 @@
-package com.tenday.core.common
+package com.tenday.core.common.extensions
 
 enum class JobPosition(val position: Int) {
     파트장(1), 대리(2), 팀장(3), 사원(4),
