+++ core/common/src/main/java/com/tenday/core/common/extensions/ProfileCode.kt
@@ -1,4 +1,4 @@
-package com.tenday.core.common
+package com.tenday.core.common.extensions
 
 enum class ProfileCode {
     F_A,
